refactor: decouple demos from expression_evaluation
